renaming "pastis diploid" to "pastis poisson"
