+++ pastis/optimization/utils_poisson.py
